+++ src/as_display/mod.rs
@@ -1,0 +1,7 @@
+mod as_debug_wrapper;
+mod debug_as_display;
+mod debug_wrapper;
+
+pub use as_debug_wrapper::AsDebugWrapper;
+pub use debug_as_display::DebugAsDisplay;
+pub use debug_wrapper::DebugWrapper;